Moved tile group methods to TileGroupFactory and added an instance variable of type TileGroupFactory to AddTileInteractor.
